third_party/blink/web_tests/platform/linux/inspector-protocol/css/css-disable-local-fonts-expected.txt

Test css.setLocalFontsEnabled method.

Running test: testElementFontsWithLocalFontsEnabled
Expected p font: "Arial local" on Win/Mac and "DejaVu Sans" on Linux
Actual p font: "DejaVu Sans local"

Running test: testElementFontsWithLocalFontsDisabled
Expected p font: "Ahem network"
Actual p font: "Ahem (Fontations) network"
